IO: add DatasetBase.py
This can avoid import circular dependency problem between IO.py and *Dataset.py
