+++ onmt/io/IO.py
@@ -1,19 +1,16 @@
 # -*- coding: utf-8 -*-
 
 from collections import Counter, defaultdict
-from itertools import chain, count
+from itertools import count
 
 import torch
 import torchtext.data
 import torchtext.vocab
 
-from onmt.Utils import aeq
-
-
-PAD_WORD = '<blank>'
-UNK = 0
-BOS_WORD = '<s>'
-EOS_WORD = '</s>'
+from onmt.io.DatasetBase import PAD_WORD, BOS_WORD, EOS_WORD
+from onmt.io.TextDataset import TextDataset
+from onmt.io.ImageDataset import ImageDataset
+from onmt.io.AudioDataset import AudioDataset
 
 
 def _getstate(self):
@@ -42,9 +39,6 @@ def get_fields(data_type, n_src_features, n_tgt_features):
         A dictionary whose keys are strings and whose values are the
         corresponding Field objects.
     """
-    # Hide this import inside to avoid circular dependency problem.
-    from onmt.io import TextDataset, ImageDataset, AudioDataset
-
     if data_type == 'text':
         return TextDataset.get_fields(n_src_features, n_tgt_features)
     elif data_type == 'img':
@@ -109,9 +103,6 @@ def get_num_features(data_type, corpus_file, side):
         number of features on `side`.
     """
     assert side in ["src", "tgt"]
-
-    # Hide this import inside to avoid circular dependency problem.
-    from onmt.io import TextDataset
 
     # For target side, all 'data_type' are in form of text.
     if side == 'tgt':
@@ -186,9 +177,6 @@ def build_dataset(fields, data_type, src_path, tgt_path, src_dir=None,
                   dynamic_dict=True, sample_rate=0,
                   window_size=0, window_stride=0, window=None,
                   normalize_audio=True, use_filter_pred=True):
-
-    # Hide this import inside to avoid circular dependency problem.
-    from onmt.io import TextDataset, ImageDataset, AudioDataset
 
     # Build src/tgt examples iterator from corpus files, also extract
     # number of features.
@@ -279,9 +267,6 @@ def _make_examples_nfeats_tpl(data_type, src_path, src_dir,
     on source side for different 'data_type'.
     """
 
-    # Hide this import inside to avoid circular dependency problem.
-    from onmt.io import TextDataset, ImageDataset, AudioDataset
-
     if data_type == 'text':
         src_examples_iter, num_src_feats = \
             TextDataset.make_text_examples_nfeats_tpl(
@@ -314,112 +299,3 @@ class OrderedIterator(torchtext.data.Iterator):
             for b in torchtext.data.batch(self.data(), self.batch_size,
                                           self.batch_size_fn):
                 self.batches.append(sorted(b, key=self.sort_key))
-
-
-class ONMTDatasetBase(torchtext.data.Dataset):
-    """
-    A dataset basically supports iteration over all the examples
-    it contains. We currently have 3 datasets inheriting this base
-    for 3 types of corpus respectively: "text", "img", "audio".
-
-    Internally it initializes an `torchtext.data.Dataset` object with
-    the following attributes:
-
-     `examples`: a sequence of `torchtext.data.Example` objects.
-     `fields`: a dictionary associating str keys with `torchtext.data.Field`
-        objects, and not necessarily having the same keys as the input fields.
-    """
-    def __getstate__(self):
-        return self.__dict__
-
-    def __setstate__(self, d):
-        self.__dict__.update(d)
-
-    def __reduce_ex__(self, proto):
-        "This is a hack. Something is broken with torch pickle."
-        return super(ONMTDatasetBase, self).__reduce_ex__()
-
-    def collapse_copy_scores(self, scores, batch, tgt_vocab):
-        """
-        Given scores from an expanded dictionary
-        corresponeding to a batch, sums together copies,
-        with a dictionary word when it is ambigious.
-        """
-        offset = len(tgt_vocab)
-        for b in range(batch.batch_size):
-            index = batch.indices.data[b]
-            src_vocab = self.src_vocabs[index]
-            for i in range(1, len(src_vocab)):
-                sw = src_vocab.itos[i]
-                ti = tgt_vocab.stoi[sw]
-                if ti != 0:
-                    scores[:, b, ti] += scores[:, b, offset + i]
-                    scores[:, b, offset + i].fill_(1e-20)
-        return scores
-
-    @staticmethod
-    def coalesce_datasets(datasets):
-        """Coalesce all dataset instances. """
-        final = datasets[0]
-        for d in datasets[1:]:
-            # `src_vocabs` is a list of `torchtext.vocab.Vocab`.
-            # Each sentence transforms into on Vocab.
-            # Coalesce them into one big list.
-            final.src_vocabs += d.src_vocabs
-
-            # All datasets have same number of features.
-            aeq(final.n_src_feats, d.n_src_feats)
-            aeq(final.n_tgt_feats, d.n_tgt_feats)
-
-            # `examples` is a list of `torchtext.data.Example`.
-            # Coalesce them into one big list.
-            final.examples += d.examples
-
-            # All datasets have same fields, no need to update.
-
-        return final
-
-    # Below are helper functions for intra-class use only.
-
-    def _join_dicts(self, *args):
-        """
-        Args:
-            dictionaries with disjoint keys.
-
-        Returns:
-            a single dictionary that has the union of these keys.
-        """
-        return dict(chain(*[d.items() for d in args]))
-
-    def _peek(self, seq):
-        """
-        Args:
-            seq: an iterator.
-
-        Returns:
-            the first thing returned by calling next() on the iterator
-            and an iterator created by re-chaining that value to the beginning
-            of the iterator.
-        """
-        first = next(seq)
-        return first, chain([first], seq)
-
-    def _construct_example_fromlist(self, data, fields):
-        """
-        Args:
-            data: the data to be set as the value of the attributes of
-                the to-be-created `Example`, associating with respective
-                `Field` objects with same key.
-            fields: a dict of `torchtext.data.Field` objects. The keys
-                are attributes of the to-be-created `Example`.
-
-        Returns:
-            the created `Example` object.
-        """
-        ex = torchtext.data.Example()
-        for (name, field), val in zip(fields, data):
-            if field is not None:
-                setattr(ex, name, field.preprocess(val))
-            else:
-                setattr(ex, name, val)
-        return ex
