+++ onmt/io/TextDataset.py
@@ -10,8 +10,7 @@ import torch
 import torchtext
 
 from onmt.Utils import aeq
-from onmt.io.IO import ONMTDatasetBase, \
-                        PAD_WORD, BOS_WORD, EOS_WORD
+from onmt.io.DatasetBase import ONMTDatasetBase, PAD_WORD, BOS_WORD, EOS_WORD
 
 
 class TextDataset(ONMTDatasetBase):
