+++ onmt/io/__init__.py
@@ -1,9 +1,10 @@
-from onmt.io.IO import PAD_WORD, BOS_WORD, EOS_WORD, UNK, ONMTDatasetBase, \
-                       collect_feature_vocabs, make_features, \
+from onmt.io.IO import collect_feature_vocabs, make_features, \
                        collect_features, get_num_features, \
                        load_fields_from_vocab, get_fields, \
                        save_fields_to_vocab, build_dataset, \
                        build_vocab, merge_vocabs, OrderedIterator
+from onmt.io.DatasetBase import ONMTDatasetBase, PAD_WORD, BOS_WORD, \
+                                EOS_WORD, UNK
 from onmt.io.TextDataset import TextDataset, ShardedTextCorpusIterator
 from onmt.io.ImageDataset import ImageDataset
 from onmt.io.AudioDataset import AudioDataset
